push

+++ Project collect the dudes/Serialization/InternalSettings.cs
@@ -31,14 +31,14 @@
         #endregion
         #region Generators
         #region Units
-        public static int baseAttributeValue => 1;
-        public static int capacityAttributeValue => 20;
-        public static int baseAttributePoints => 5;
-        public static int oneStarAttributePointsMaximum => 10;
-        public static int twoStarAttributePointsMaximum => 20;
-        public static int threeStarAttributePointsMaximum => 35;
-        public static int fourStarAttributePointsMaximum => 45;
-        public static int fiveStarAttributePointsMaximum => 55;
+        public static byte baseAttributeValue => 1;
+        public static byte capacityAttributeValue => 20;
+        public static byte baseAttributePoints => 5;
+        public static byte oneStarAttributePointsMaximum => 10;
+        public static byte twoStarAttributePointsMaximum => 20;
+        public static byte threeStarAttributePointsMaximum => 35;
+        public static byte fourStarAttributePointsMaximum => 45;
+        public static byte fiveStarAttributePointsMaximum => 55;
 
         public static float healthPerStrength => 10;
         public static float healthRegenPerStrength => 0.1f;
@@ -47,7 +47,7 @@
         public static float manaRegenPerIntelligence => 0.5f;
         public static float staminaPerAgility => 10;
         public static float staminaRegenPerAgility  => 1;
-        public static float speedPerAgility => 1;
+        public static byte speedPerAgility => 1;
         public static float physicalResitancePerStrength => 2;
         public static float magicResistancePerIntelligence => 2;
         public static float hitChancePerAgility => 2.5f;
